requirements.txt
streamlit>=1.28.1
controlflow>=0.12.0
pymongo>=4.5.0
pandas>=2.1.1
python-dotenv>=1.0.0
openai>=1.55.3
sentence-transformers>=2.2.2
numpy>=1.24.3
pydantic>=2.4.2
structlog>=23.2.0
